Improve search functionality and add grading filter
The update modifies the search and filter functionality in the table component to cater to more specific cases. The filter now supports different search keys like 'id' and 'grade' with an improved regex for extracting numbers. Filtering for grading is now available, supporting a range of values. Redundancies in setting the filter value have been removed in the filter component and delay times have been adjusted.