Update filter feature and improve validation on form fields
Added SearchService to handle filter queries in the table component. Modified store to include an action and reducer for updating trainees, as well as changes to training service to handle updates. Improved validation on form fields in the details component and added the ability to show error messages when fields are not filled properly.