Add new charts component and enhance trainee details view
A new charts component has been added to the 'analysis-container' along with its corresponding tests, styles, and HTML template. Additionally, some changes have been made to increase the clarity of the trainee details view. The 'viewButton$' is now disabled instead of hidden when not in use, and further checks have been added to ensure the trainee form values are correctly updated.